Eth package refactor
