+++ packages/eth/package.json
@@ -33,7 +33,7 @@
         }
     ],
     "dependencies": {
-        "pocket-js": "../core/",
+        "pocket-js-core": "../core/",
         "pocket-js-web3-provider": "../web3-provider/",
         "rxjs": "^6.4.0",
         "web3": "^1.0.0-beta.51",
