+++ packages/eth/src/index.js
@@ -1,0 +1,1 @@
+module.exports = Object.assign(module.exports, require('./eth'), require('./eth-contract'));
