+++ packages/eth/src/rpc/eth.js
@@ -4,9 +4,8 @@
  */
 // Constants
 const BlockTag = require('../models/blocktag.js').BlockTag;
-const RpcUtils = require('./rpcUtils.js');
-const EthAccounts = require('web3-eth-accounts').Accounts;
-const PocketProvider = require('pocket-js-web3-provider').PocketProvider;
+const RpcUtils = require('./utils');
+const PocketProvider = require('pocket-js-web3-provider');
 const Web3 = require('web3');
 const Web3Utils = require('web3-utils');
 const ETH_RPC_METHODS = Object.freeze({
