+++ packages/eth/tests/mocha.opts
@@ -1,0 +1,1 @@
+--timeout 0
